Update patch set 2
Patch Set 2:

Our policy is that 'master' is where new features and fixes go; in general we don't expect to support old versions.

However, because several API-breaking changes are planned for 'master' in the near future, we wanted to label (with a branch) the commit before the AArch32 port. This patch tests our process for back ports, to ensure that we _can_ do it if we have to.

Patch-set: 2