Update patch set 1
Patch Set 1:

(2 comments)

This implementation suffers from one significant issue - all the components of an instruction (destination register, label, etc.) are processed in isolation, but interpreting them sometimes requires being aware of the whole context. Examples have been given inline.

We discussed several solutions: One of them is to make all instruction methods (e.g. void b(Condition cond, EncodingSize size, Label* label) ) virtual. Unfortunately, this approach does not scale well because users will have to manually override all instruction variants (such as more than 20 load and store types). An alternative is to define a mechanism to query the properties of an instruction; in this case users may subclass DisassemblerStream to keep a pointer to the current instruction, and use the new interface to obtain the necessary information. The final option is to use the same interface as the AArch64 disassembler, which implies a significant change of the current implementation.

Patch-set: 1