Update patch set 1
Patch Set 1:

I discussed this with Jacob, and he was not keen on having `-inl.h` files. Other options are to move the code to the bottom of `.h` files, to inline the code in the classes, or obviously to keep things as they are.

I am looking for opinions here.

Patch-set: 1