Update patch set 4
Patch Set 4:

I'm unsure about this. While this works, it's surprising to have the simulator gate a feature (VIXL_HAS_RUNTIME_SUPPORT) which is not only related to the simulator. You are allowed to use `CallRuntime` when running natively with the simulator turned off.

It happens to work natively because we do not have VIXL_INCLUDE_SIMULATOR guards around simulator-aarch64.h, but it is also surprising that do not.

Patch-set: 4
Label: Code-Review=0