Allow specifying whether an instruction should set flags with an `enum`
This patch allow the macro-assembler user to call instructions which can
set flags with an argument.

That is instead of calling "add(rd, ...)" or "adds(rd, ...)" you can call
"add(flags, rd, ...)" with flags which can take the values:
 - SetFlags
 - LeaveFlags
 - DontCare

It will call the right function (add or adds) depending on flags and the context.

Change-Id: I46d07d3b2989522d16d6432a469dc4b796682cbd