Update patch set 4
Patch Set 4:

Your assumption that there will be no wrong guesses is tenuous at best, but let's say you are right. Well, in that case you transfer the risk to the significantly more complicated size calculation - a very easy mistake to make is to miss a star in the expression, which no amount of checking performed by the compiler will be able to catch (and ends up being pretty much the same issue as the one supposedly in my original proposal). At the same time the improvement to DisassembleT32Buffer() amounts to replacing "+=" with "=", or in other words - minimal.

So, why are you pushing for an approach which suffers from exactly the same, in your own words, "major problem", with uncertain improvements, and different type of risks? If some comments are what you require, just say so.

Patch-set: 4
Label: Code-Review=0