Update patch set 4
Patch Set 4:

I can't think of good reasons to have Disassembler{A32,T32}Buffer(...) methods use uint32_t* and uint16_t* respectively. This can be confusing when compared to VIXL64 where we use Instruction* which is 1 byte aligned.

Anton, is it a lot of work to change those methods too with this patch? It could be done later too but I don't believe there are a lot instance of them around.

Patch-set: 4
Label: Code-Review=0