Update patch set 6
Patch Set 6: Code-Review-2

I understand less and less this patch. It looks like it allow more and more "bad" use of the library instead of promoting "good" use.
Being 16 bits aligned for T32 and 32 bits aligned is a requirement of ARM CPUs. Even if we could understand to be able to disassemble a non aligned buffer, I think it's a bad idea. Of course using a uint16_t* won't enforce that the buffer is aligned but, it would help.

T32 and A32 prototypes are not consistent (for DecodeAT).
As you insist to return an int, at least you can return "-1" if you can't decode an instruction instead of returning a valid value which is misleading.

Patch-set: 6
Label: Code-Review=-2