Update patch set 5
Patch Set 5: Code-Review-1

I'm not sure it's a good idea to remove the Assembler wrappers. The buffer is a property of the assembler and for usual cases, it could be totally hidden. The wrappers are really useful to reduce the complexity of the calls.

Patch-set: 5
Reviewer: Vincent Belliard <1000221@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=-1