Improve constant move helper to use PC-relative instructions
If the constant being moved in to the register is a pointer within 4GB of the assembler's cursor then you can use adrp and adr to load it in an instruction or two.
If your pointer is within 1MB, you can load the pointer in one instruction. If your pointer is 4KB page aligned, you can load the pointer in one instruction.
If the pointer isn't 4KB aligned, you can load the pointer in two instructions.

Change-Id: I55ecab3b0f9e5a3f9fc5ea50361b1d371f3c7163