Update patch set 2
Patch Set 2:

(7 comments)

Thanks!

There are test failures but they appear to be formatting-related. You can run these yourself using tools/test.py; if it's set up correctly it will run the linter as part of that. (README.md should explain how to set it up, but if not, please let us know and we'll fix it!)

Also, this needs some tests. Once the design is finalised, please add a few tests to test/aarch64/test-assembler-aarch64.cc. For example, the tests should verify that addresses are correctly loaded in each of the ranges handled by adrp and/or adr, and if a mechanism like MoveImmediateHelper is used, the tests should also verify that the correct number of instructions is reported.

Patch-set: 2