Ensure `MacroAssembler::Nop` generates a single `nop`.
The `MacroAssembler` is allowed to emit any code that is functionally
equivalent to what is requested by the user, and is expected to
generate 'good' code.  The user can expect calls such as
`masm.Add(r0, r0, 0)` to be optimized away (think that `0` may be the
value of a variable).
Following this we *could* generate no code for calls to `Nop()`. But
`Nop()` will likely not be used by mistake, so we ensure it generates
exactly one `nop` instruction. This is useful for tests where we want
to generate 'some' code.

Change-Id: Iecf31b5cfd151a77b1633ea5158ff54fef9f8908