AArch32: Improve tests related to pool emissions while in a delegate.
The main goal of the new version is to not be dependent on the behaviour of the
`MacroAssembler` regarding the margin it uses to generate literal and veneer
pool. So for example the new tests do not assume that

    masm.Mov(r0, r1);

will only check whether pools are required in the next 2 bytes.

The tests themselves were validated by partially reverting:

    commit dcffac4d0a5a586d3c14971e33bd28fc848bc148
    Author: Vincent Belliard <vincent.belliard@arm.com>
    Date:   Wed Oct 19 11:31:20 2016 -0700

        Fix veneer/pool emission while in a Delegate.

        Change-Id: I76ff3e224f5e85a58a53865c62536d97e73b6730

and

    commit e42218c6ca969b7d4032da978fb05f06641df100
    Author: Vincent Belliard <vincent.belliard@arm.com>
    Date:   Wed Oct 19 13:24:28 2016 -0700

        Make bind and place more consistent.

        Change-Id: I1743511e3c6f995f05cded38347a015c198b1fb9

to restore the problematic behaviours.

Change-Id: Ibfde6bcfc8778ca375f2dd48e0216b35170318df