Update patch set 15
Patch Set 15:

> I don't see how this patch can work. The two major problems seem to
 > be:
 > - If you allocate only an Assembler object, it won't allow
 > assembler instructions.
 > - When a macro assembler instruction is called, the scope no longer
 > allow to call assembler instructions (and that's what is done by
 > the macro assembler instruction).

It passes all the tests, and follows the guidelines in `doc/design/code-generation.md`.
The assembler is enabled whenever a `CodeBufferCheckScope` is created; previously we said it 'acquires' the buffer. You always need to create a scope before generating code to ensure you have space in your buffer.

Patch-set: 15
Label: Code-Review=0