[aarch32] Cleanup *_pool_in_delegate tests
The veneer_pool_in_delegate and literal_pool_in_delegate tests were written to
make sure that generating a pool from inside a MacroAssembler delegate was
working. However, we no longer allow this and block the pool before delegates
can be called. Instead, we generate pools before handling the instruction, if we
need to.

So cleanup and rename those test to avoid confusion. They now just test that a
macro-instruction does generate the pools first. Additionnaly, the veneer was
changed so the branches would be taken, to make sure the veneers are correct.

Change-Id: Idc60d8b874aa2602711f8329d34e3f48f4ae0bb0