[aarch64] Do not mix the Assembler and MacroAssembler in tests
Cleanup all of vixl64's test to stop use ExactAssemblyScope instead of
CodeBufferCheckScope. The latter allows for mixing the assembler and
macro-assembler.

Fixing the disassembler tests were not completely obvious. We used to
have a different SETUP macro for the Assembler and MacroAssembler. It's
much simpler to have a single SETUP macro that always creates a
MacroAssembler and then use COMPARE or COMPARE_MACRO to define the
tests. The former allows the Assembler with an ExactAssemblyScope while
the latter does not.

Change-Id: Ic37c25742b77767f257f88e06c7e4106ea55425f