Refactor scope implementation.
This is a redesigning of the internal scope implementation, allowing for easier
extension in the future and better understanding of "who does what".

We had a class hierarchy based design where scopes would inherit their
functionnality from other scopes. However, has bugs have shown, it's hard to
make sure all scopes behave the same way when we do this. If we are introducing
a new scope, we have to make sure the semantics of "Open", "Close", the
constructor and destructor are the same.

With this patch we have a common internal "Scope" class which implements those
semantics but defers to calling internal "DoOpen" and "DoClose" methods which do
the real work. When defining a new scope, we have to subclass it and define
those internal methods.

Not using a class hierarchy means we are duplicating code in those "DoOpen" and
"DoClose" methods. However, it makes very easy to understand what a scope
actually does, we no longer have to look up what the base scope does.

Change-Id: I2685c7bc048655e23551e792e6ce14a32c4708f4