Update patch set 1
Patch Set 1:

(7 comments)

I did not spend much time reviewing.

When I implemented this I knew that the code was a bit tough, but my point of view was that once working, we would touch scopes as little as possible. When I worked on it I remember pushing for deduplication, because issues had creeped in with the duplication (can't remember the details).
So I am a little careful when approaching a full refactoring like this. (Although if it passes all the tests it should be fine.)

This increases the scopes file from ~300 to ~500 lines. Are we fixing some bugs (new tests?) or improving the current status significantly?
I can see how the new implementation would help implementing scopes handling more features, while a hierarchy of subclasses may make it awkward; but isn't it enough today? Or are we planning for tomorrow?

Patch-set: 1
Reviewer: Alexandre Rames <1000284@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0