Clear floating-point flags in SETUP.
Some tests assume that the flags are cleared before running them. For example,
vcmp tests assume that all the flags that are not affected by the vcmp
instruction will be zero, which might not be the case if an inexact conversion
has happened, for instance. Clearing the flags makes the tests work reliably.

Change-Id: I02b2f367db5ce8219a0028d99db59766f0355e46