Give CPUFeatures::kNone a stable value.
This makes the simulation pseudo-instruction mechanism a bit more stable across
VIXL versions.

We still don't guarantee that simulation pseusdo-instructions generated with the
MacroAssembler will work with a Simulator from a different version of VIXL, but
this patch has no real cost for VIXL, and might make things easier for someone.

Change-Id: Idd92328801a88f54a8d30db86304746a22ccb1df