Add option to regroup tests execution at the end.
Defer the execution of all the tests at the end of the pipeline.

Instead of having builds and tests runs interleaved, all the builds are
done first then all the tests are run: this speeds up the tests
execution by about 50%.

Refactored the content of threaded_tests into some helper classes.

Change-Id: I2842f9e8657f6b955f8bc4fcc5654d96aeec9648