Introduce IntegerOperand.
For signed immediate fields, we'd like users to be able to provide
either signed values or unsigned (typically hexadecimal)
representations. Typically, we want to accept values in the range
[ INT<N>_MIN, UINT<N>_MAX ] for elements with N bits.

SVE has a lot of signed immediate fields, so this patch introduces a new
IntegerOperand class to make the range checks convenient.

Until now, we have achieved this by simply accepting uint64_t, since C++
will sign-extend smaller negative types. However, this does not allow us
to reject large values (such as an attempt to move 0xffffffffffffffff
into a byte-sized element). In addition, each function with such an
argument then has to duplicate the range check, which generally requires
a comment because it has some subtle behaviours.

Change-Id: If2b9200f3094f01cc8a889869980aaa6b72c3103